+++ src/client/components/article/pages/404.md
@@ -1,1 +1,11 @@
-# ikke fundet
+![laesekompas-logo](https://user-images.githubusercontent.com/36037828/48554601-20c60700-e8df-11e8-80d7-b2dc035a416a.png)
+
+## 404
+
+#### Siden blev ikke fundet
+
+Siden du prøver at finde, findes desværre ikke.
+
+---
+
+[Tilbage til forsiden](/)
