+++ src/client/components/article/pages/about.md
@@ -1,5 +1,32 @@
-![laesekompas-logo](https://user-images.githubusercontent.com/36037828/48554601-20c60700-e8df-11e8-80d7-b2dc035a416a.png)
+# Om Læsekompasset
 
-### Om læsekompas.dk
+### Læsekompasset er dit online inspirationsværktøj til at finde ny læsning
 
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent sed aliquam ante. Vivamus a interdum felis. Fusce tincidunt arcu arcu, ut porta quam pharetra ut. Donec vitae metus ut purus interdum molestie. Nunc feugiat interdum consequat. Vestibulum euismod nisi lectus. Proin commodo orci id metus euismod posuere.
+Læsekompasset er bibliotekernes inspirationsværktøj til voksne læsere af skønlitteratur. Alle kan anvende Læsekompasset, men det kræver biblioteks-login, hvis du vil låne bøger igennem websitet.
+
+Læsekompasset indeholder et udvalg af skønlitteratur for voksne, der er tilgængeligt til hjemlån eller som e-lån fra folkebiblioteker. Mængden af værker øges over tid og Læsekompasset sigter mod at tilbyde en bred vifte af litteratur. Alle værker indekseres med oplevelsesorienterede metadata for at gøre det muligt for brugere at navigere i litteratur på en ny måde. Der er mulighed for at fremfinde værker og bestille dem direkte i Læsekompasset.
+
+Læsekompasset giver også brugere mulighed for at oprette og dele lister med bøger.
+
+### Betaversion af Læsekompasset
+
+Læsekompasset er lanceret som en betaversion. Dette betyder, at vi fortsat tester og justerer. Den version af Læsekompasset, du oplever er ikke helt færdigudviklet og du vil opleve at løsningen udvikler sig over tid.
+
+### Hvem står bag?
+
+Læsekompasset er skabt i samarbejde med de danske folkebiblioteker og drives af DBC.
+
+### Brugergeneret indhold
+
+Følgende regler gælder for indlæg eller indhold, som du og andre brugere selv lægger på websitet, i det omfang vi åbner mulighed for dette. Du er ansvarlig for, at dine indlæg er i overensstemmelse med den til enhver tid gældende lovgivning, herunder at dine indlæg ikke krænker tredjemands rettigheder, eksempelvis ophavsret til tekst og billeder. Du giver DBC og Læsekompasset ret til at mangfoldiggøre egne fotos, som du uploader til Læsekompasset, at gøre dem tilgængelige for tredjemand og at lave afledte værker.
+
+For alle informationer og alt materiale, du leverer til websitet gælder, at du leverer det vederlagsfrit og til fri brug.
+
+Læsekompasset kan ikke kontrollere og er ikke ansvarlig for indhold, som brugerne leverer, og for at du som bruger kan risikere at blive udsat for indhold, som du finder anstødeligt, fejlagtigt eller på anden måde kritisabelt. Du anerkender, at vi som udbyder ikke forhåndsgodkender brugernes indhold, men at vi har retten til uden nærmere begrundelse at fjerne enhver form for indhold, som offentliggøres via Læsekompasset, hvis det er i modstrid med gældende lovgivning eller ordlyden af de her nævnte betingelser.
+
+### Kontakt Læsekompasset:
+
+DBC as  
+Tempovej 7-11  
+2750 Ballerup  
+dbc@dbc.dk
